--- a/src/astro/databases/postgres.py
+++ b/src/astro/databases/postgres.py
@@ -6,7 +6,6 @@ from typing import Dict, List
 import pandas as pd
 import sqlalchemy
 from airflow.providers.postgres.hooks.postgres import PostgresHook
-from pandas.io.sql import SQLDatabase
 from psycopg2 import sql as postgres_sql
 
 from astro.constants import DEFAULT_CHUNK_SIZE, LoadExistStrategy, MergeConflictStrategy
@@ -73,18 +72,6 @@ class PostgresDatabase(BaseDatabase):
         :param if_exists: Strategy to be used in case the target table already exists.
         :param chunk_size: Specify the number of rows in each batch to be written at a time.
         """
-        if target_table.metadata and target_table.metadata.schema:
-            self.create_schema_if_needed(target_table.metadata.schema)
-
-        # create an empty table
-        db = SQLDatabase(engine=self.sqlalchemy_engine)
-        db.prep_table(
-            source_dataframe,
-            target_table.name.lower(),
-            schema=target_table.metadata.schema,
-            if_exists=if_exists,
-            index=False,
-        )
 
         output_buffer = io.StringIO()
         source_dataframe.to_csv(output_buffer, sep=",", header=True, index=False)