--- a/tests/benchmark/dags/evaluate_load_file.py
+++ b/tests/benchmark/dags/evaluate_load_file.py
@@ -63,7 +63,7 @@ def create_dag(database_name, table_args, dataset):
             output_table=table,
             chunk_size=chunk_size,
         )
-        aql.truncate(my_table)
+        aql.cleanup([my_table])
 
         # Todo: Check is broken so the following code is commented out, uncomment when fixed
         # aggregate_check(