--- a/tests/sql/operators/test_load_file.py
+++ b/tests/sql/operators/test_load_file.py
@@ -641,38 +641,6 @@ def test_aql_nested_ndjson_file_to_bigquery_explicit_illegal_sep_params(
     assert "payload_size" in df.columns
 
 
-@pytest.mark.parametrize(
-    "database_table_fixture",
-    [
-        {
-            "database": Database.POSTGRES,
-        },
-    ],
-    indirect=True,
-    ids=["postgresql"],
-)
-def test_aql_multilevel_nested_ndjson_file_default_params(
-    sample_dag, database_table_fixture, caplog
-):
-    """
-    Test the flattening of multilevel level nested ndjson, with default '_'.
-    Expected to fail since we do not support flattening of multilevel ndjson.
-    """
-    _, test_table = database_table_fixture
-
-    with pytest.raises(BackfillUnfinished):
-        with sample_dag:
-            load_file(
-                input_file=File(
-                    path=str(CWD) + "/../../data/github_multi_level_nested.ndjson"
-                ),
-                output_table=test_table,
-            )
-        test_utils.run_dag(sample_dag)
-    expected_error = "can't adapt type 'dict"
-    assert expected_error in caplog.text
-
-
 def test_populate_table_metadata(sample_dag):
     """
     Test default populating of table fields in load_fil op.