# Description
## What is the current behavior?
Currently the postgres load numbers are far too slow for any production use-case. We need to find optimizations to bring the 10GB loading time to under 5 minutes for customers to reasonably use this feature.


Previous numbers:
| database                | dataset    | total_time   | memory_rss   | cpu_time_user   | cpu_time_system   |
|:------------------------|:-----------|:-------------|:-------------|:----------------|:------------------|
| postgres_conn_benchmark | few_kb     | 494.42ms     | 36.06MB      | 570.0ms         | 50.0ms            |
| postgres_conn_benchmark | ten_kb     | 689.05ms     | 42.96MB      | 540.0ms         | 40.0ms            |
| postgres_conn_benchmark | hundred_kb | 580.7ms      | 36.43MB      | 570.0ms         | 50.0ms            |
| postgres_conn_benchmark | ten_mb     | 44.67s       | 1.38GB       | 31.03s          | 4.03s             |
| postgres_conn_benchmark | one_gb     | 5.56min      | 62.5MB       | 14.07s          | 1.14s             |
| postgres_conn_benchmark | five_gb    | 24.44min     | 78.15MB      | 1.34min         | 5.73s             |
| postgres_conn_benchmark | ten_gb     | 45.64min     | 61.71MB      | 2.37min         | 11.48s            |

closes: [#428](https://github.com/astronomer/astro-sdk/issues/428)
<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->


## What is the new behavior?
In this PR we found two primary optimizations that have led to significant speed up

The first optimization is to not directly place a smart_open into the `read` function of pandas. There have been multiple reports of pandas and smart_open working very slow when used together. Instead, we read the `smart_open` file into an IO buffer (either StringIO or BytesIO). This also has the benefit that we do not need to write to disc if there is enough memory to hold the object.

The second optimization is postgres specific. The `pandas.to_sql` function uses the HIGHLY suboptimal `INSERT` statement for postgres. This is 10X slower than the `COPY` command. By saving objects as CSV buffers, we can use COPY to load the data into postgres as a stream instead of one value at a time.

Here are the new numbers with the optimizations:

| database   | dataset    | total_time   | memory_rss   | cpu_time_user   | cpu_time_system   |
|:-----------|:-----------|:-------------|:-------------|:----------------|:------------------|
| postgres   | ten_kb     | 432.93ms     | 43.19MB      | 540.0ms         | 60.0ms            |
| postgres   | hundred_kb | 417.2ms      | 31.34MB      | 560.0ms         | 40.0ms            |
| postgres   | ten_mb     | 2.21s        | 79.91MB      | 1.76s           | 170.0ms           |
| postgres   | one_gb     | 13.9s        | 35.43MB      | 7.97s           | 5.76s             |
| postgres   | five_gb    | 50.23s       | 43.27MB      | 26.64s          | 18.7s             |
| postgres   | ten_gb     | 3.11min      | 43.29MB      | 1.36min         | 1.48min           |

## Does this introduce a breaking change?

No

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary