@dimberman 2022-07-13
@tatiana can you point me to where this work is being done (re: allowing users to define their own schema)?