@tatiana 2022-07-18
@dimberman Sorry for the delay!
> @tatiana can you point me to where this work is being done (re: allowing users to define their schema)?
I recently added it to the base class, so it should work for Postgres as well:
https://github.com/astronomer/astro-sdk/blob/main/src/astro/databases/base.py#L279-L282
https://github.com/astronomer/astro-sdk/blob/main/src/astro/databases/base.py#L204
https://github.com/astronomer/astro-sdk/blob/main/src/astro/databases/base.py#L167